+++ SIAL_bqa_incomplete/MOE_itl_events.sql
@@ -9,7 +9,7 @@ INPUT: [IDI_CLEAN].[moe_clean].[tec_it_learner]
 OUTPUT: {schemaname}.SIAL_MOE_ITL_events
 
 DEPENDENCIES: 
-{schemaname}.[inflation_index] table must exist
+{schemaname}.[moe_itl_fund_rate] table must exist
 
 NOTES: 
 Reference period start: 
@@ -21,8 +21,8 @@ People undergoing workplace-based training activity which is eligible for fundin
 Analysis Unit: 
 Training activity of individuals in programmes administered by Industry Training Organisation in each training fund in a calendar year. 
 
-Cost is selected to be fixed at 2000 for a full course as an estimate based in the year ending 2015, until  actual cost per course from MOE
-MoE uses fixed cost for each course - awaiting fixed cost from MoE, till then use fixed figure of 2000
+Cost is taken from fixed cost per year by the fund code from MOE - given by David Earle
+
 Each row is per course for each snz_uid
 
 Note that from 2011 onwards there is a change of management system
@@ -32,6 +32,8 @@ AUTHOR: Wen Jhe Lee
 DATE: 21 April 2017
 
 HISTORY: 
+v2 - updated with actual fund cost from MOE -
+v1 created 21 Apr 2017
 
 
 *********************************************************************************************************/
@@ -49,20 +51,24 @@ select snz_uid,
 		moe_itl_ito_edumis_id_code as event_type_2, -- MOE ID of Industry Training Provider
 		moe_itl_nqf_level_code as event_type_3, -- The NZQF level of the training programme
 		moe_itl_nzsced_detail_text as event_type_4, -- The NZSCED detail of programme
-		round(sum(cost2),2)  as cost
+		round(sum(cost_sum),2)  as cost
 		from  
 					(select * ,
-						(moe_itl_sum_units_consumed_nbr*2000)*b.adj  as cost2 
-						from [IDI_CLEAN].[moe_clean].[tec_it_learner]
-					left join (
-						select value/(select value from {schemaname}.[inflation_index] where inflation_type='CPI' and quarter='2015Q4') as adj ,
-						start_date as sd,
-						end_date as ed 
-						from  {schemaname}.[inflation_index] where inflation_type='CPI') b
-					on moe_itl_start_date between sd and ed ) z 
+						(moe_itl_sum_units_consumed_nbr* moerate.rate) as cost_sum 
+						from [IDI_CLEAN].[moe_clean].[tec_it_learner] tec_learner
+					inner join (
+						select rate ,
+						start_date as sd_rate,
+						end_date as ed_rate,
+						moe_itl_fund_code as fund
+						from  {schemaname}.[moe_itl_fund_rate] ) moerate -- Table derived from MOE fixed cost by year
+					on 
+					 tec_learner.moe_itl_fund_code=moerate.fund and
+					 tec_learner.moe_itl_start_date between sd_rate and ed_rate ) summary
 					where moe_itl_start_date!='' 
 
 		group by snz_uid,moe_itl_start_date,moe_itl_end_date,moe_itl_fund_code,moe_itl_ito_edumis_id_code,	moe_itl_nqf_level_code, moe_itl_nzsced_detail_text
+
 
 )
 ;