+++ CHANGELOG.md
@@ -14,6 +14,10 @@ v1
 * New feature to enable the SIAL components to be pointed to any refresh version of the IDI_Clean DB (provided that all source tables required by the SIAL views are available in that refresh)
 * IDI_Clean.moh_clean.mortality_registrations table structure has changed since the latest IDI refresh. The latest SIAL script has been modified to accomodate for this change. What this means is that if the SIAL is repointed to IDI_Clean_20161020, then the script 'MIX_mortality_events.sql' is bound to fail. You will need to obtain the earlier version of this SIAL script from Github and replace the new script with the older version.
 
-
-
+## [1.1.1] - Aug 2017
+* Business QA completed for several MOH tables
+* Bug fix for MOJ court cases table to ensure it is joining on the right ids
+* Tidy up of the ITL csv and related tables
+* Changing the MSD T1 cost table to a table about entitlements. This table no longer includes costs. Actual benefits paid can be obtained from the IRD table
+* Tidy up of the long MSD T1 script. Formats and macros have been separated out